scripts/pyproject.toml
[project]
name = "rent-my-gear-scripts"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=2.18.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.0",
]
